move SearchStatistics in its own package
