+++ src/main/java/org/ddolib/awastar/core/solver/AwAstarSolver.java
@@ -1,10 +1,10 @@
 package org.ddolib.awastar.core.solver;
 
 import org.ddolib.common.dominance.DominanceChecker;
-import org.ddolib.common.solver.SearchStatistics;
-import org.ddolib.common.solver.SearchStatus;
 import org.ddolib.common.solver.Solution;
 import org.ddolib.common.solver.Solver;
+import org.ddolib.common.solver.stat.SearchStatistics;
+import org.ddolib.common.solver.stat.SearchStatus;
 import org.ddolib.ddo.core.Decision;
 import org.ddolib.ddo.core.SubProblem;
 import org.ddolib.modeling.AwAstarModel;
@@ -154,7 +154,7 @@ public final class AwAstarSolver<T> implements Solver {
                 root.getValue() + weight * root.getLowerBound());
 
         while (!open.isEmpty()) {
-            // -- debug, stats, verbosity, stopping  ---
+            // -- debug, stat, verbosity, stopping  ---
 
             verboseMode.detailedSearchState(statistics.nbIterations(), open.size(), bestUB,
                     open.peek().getLowerBound(), statistics.gap());
@@ -169,7 +169,7 @@ public final class AwAstarSolver<T> implements Solver {
                 return new Solution(bestSolution(),
                         statistics.updateTime(System.currentTimeMillis()));
             }
-            // -- end debug, stats, verbosity, stopping  ---
+            // -- end debug, stat, verbosity, stopping  ---
 
             SubProblem<T> sub = open.poll();
             openByF.remove(sub);
