+++ src/main/java/org/ddolib/common/solver/stat/SearchStatus.java
@@ -1,4 +1,4 @@
-package org.ddolib.common.solver;
+package org.ddolib.common.solver.stat;
 
 /**
  * Status values describing the state of a search process.
