+++ src/main/java/org/ddolib/ddo/core/solver/SequentialSolver.java
@@ -1,10 +1,10 @@
 package org.ddolib.ddo.core.solver;
 
 import org.ddolib.common.dominance.DominanceChecker;
-import org.ddolib.common.solver.SearchStatistics;
-import org.ddolib.common.solver.SearchStatus;
 import org.ddolib.common.solver.Solution;
 import org.ddolib.common.solver.Solver;
+import org.ddolib.common.solver.stat.SearchStatistics;
+import org.ddolib.common.solver.stat.SearchStatus;
 import org.ddolib.ddo.core.Decision;
 import org.ddolib.ddo.core.SubProblem;
 import org.ddolib.ddo.core.cache.SimpleCache;
@@ -273,7 +273,7 @@ public final class SequentialSolver<T> implements Solver {
         statistics = statistics.updateTime(System.currentTimeMillis());
         if (bestSol.isPresent()) statistics = statistics.updateStatus(SearchStatus.OPTIMAL).updateGap(0);
         else statistics = statistics.updateStatus(SearchStatus.UNSAT);
-        
+
         return new Solution(bestSolution(), statistics);
     }
