+++ src/main/java/org/ddolib/examples/max2sat/Max2SatAcsMain.java
@@ -1,7 +1,7 @@
 package org.ddolib.examples.max2sat;
 
-import org.ddolib.common.solver.SearchStatistics;
 import org.ddolib.common.solver.Solution;
+import org.ddolib.common.solver.stat.SearchStatistics;
 import org.ddolib.modeling.AcsModel;
 import org.ddolib.modeling.Problem;
 import org.ddolib.modeling.Solvers;
