+++ src/main/java/org/ddolib/examples/maximumcoverage/MaxCoverDdoMain.java
@@ -1,23 +1,15 @@
 package org.ddolib.examples.maximumcoverage;
 
-import org.ddolib.common.solver.SearchStatistics;
 import org.ddolib.common.solver.Solution;
-import org.ddolib.ddo.core.heuristics.cluster.*;
 import org.ddolib.ddo.core.heuristics.width.FixedWidth;
 import org.ddolib.ddo.core.heuristics.width.WidthHeuristic;
-import org.ddolib.examples.smic.SMICRanking;
-import org.ddolib.examples.smic.SMICRelax;
-import org.ddolib.examples.smic.SMICState;
 import org.ddolib.modeling.DdoModel;
 import org.ddolib.modeling.Problem;
 import org.ddolib.modeling.Solvers;
 import org.ddolib.util.io.SolutionPrinter;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
-import java.util.*;
+
 /**
  * Maximum Coverage (MaxCover) problem with Ddo
  * <p>
@@ -60,7 +52,7 @@ public class MaxCoverDdoMain {
      * @throws IOException if an error occurs while loading a problem instance from a file
      */
     public static void main(String[] args) throws IOException {
-        MaxCoverProblem problem = new MaxCoverProblem(30, 30, 7,0.1,42);
+        MaxCoverProblem problem = new MaxCoverProblem(30, 30, 7, 0.1, 42);
         // MaxCoverProblem problem = new MaxCoverProblem(10, 10, 5,0.1,42);
 
         // MaxCoverProblem problem = new MaxCoverProblem("src/test/resources/MaxCover/mc_n10_m5_k3_r10_0.txt");
@@ -98,7 +90,7 @@ public class MaxCoverDdoMain {
         };
 
         Solution solution = Solvers.minimizeDdo(model, (sol, s) -> {
-            SolutionPrinter.printSolution(s,sol);
+            SolutionPrinter.printSolution(s, sol);
         });
         System.out.println();
         System.out.println(solution);
