+++ src/main/java/org/ddolib/examples/mcp/MCPAstarMain.java
@@ -1,7 +1,7 @@
 package org.ddolib.examples.mcp;
 
-import org.ddolib.common.solver.SearchStatistics;
 import org.ddolib.common.solver.Solution;
+import org.ddolib.common.solver.stat.SearchStatistics;
 import org.ddolib.modeling.Model;
 import org.ddolib.modeling.Problem;
 import org.ddolib.modeling.Solvers;
